requirements.txt
Flask==3.0.0
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
Werkzeug==3.0.1
